Unbreak the old format function, as it's used by service:check as well.
