+++ packages/apollo/src/diff/index.ts
@@ -26,10 +26,6 @@ import {
   DiffInputValue,
   DiffEnum
 } from "./ast";
-import {
-  ValidationResult,
-  ValidationErrorType
-} from "apollo-language-server/lib/engine/operations/validateOperations";
 
 export * from "./ast";
 
@@ -40,21 +36,21 @@ export interface SchemaChange {
   description: string;
 }
 
-export function format({ type, description }: ValidationResult, index: number) {
-  let color = (x: string) => x;
-  switch (type) {
-    case ValidationErrorType.FAILURE:
-      color = chalk.red;
-      break;
-    case ValidationErrorType.INVALID:
-      color = chalk.gray;
-      break;
-    case ValidationErrorType.WARNING:
-      color = chalk.yellow;
-      break;
+export const format = (change: SchemaChange) => {
+  let color = (x: string): string => x;
+  if (change.type === ChangeType.FAILURE) {
+    color = chalk.red;
   }
-  return `    [${index}] ${color(type)}    ${description}`;
-}
+  if (change.type === ChangeType.WARNING) {
+    color = chalk.yellow;
+  }
+
+  return {
+    type: color(change.type),
+    code: color(change.code),
+    description: color(change.description)
+  };
+};
 
 // A lot / most of this code is lifted from non exported members of graphql-js
 // really amazing work in there
